Revert "small fix for transfer tag, make it always 'long long', never 'int'"

+++ include/megaapi.h
@@ -3357,7 +3357,7 @@ class MegaRequest
          *
          * @return Tag of a transfer related to the request
          */
-        virtual long long getTransferTag() const;
+        virtual int getTransferTag() const;
 
         /**
          * @brief Returns the number of details related to this request
